+++ core/coroutines/test/src/commonMain/kotlin/ru/vladislavsumin/core/coroutines/test/DispatcherUtils.kt
@@ -1,4 +1,4 @@
-package ru.vladislavsumin.core.decompose.test
+package ru.vladislavsumin.core.coroutines.test
 
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
@@ -7,7 +7,6 @@ import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.setMain
 import kotlin.coroutines.ContinuationInterceptor
 
-// TODO Sumin: вынести отдельным тестовым модулем
 /**
  * Устанавливает dispatcher [TestScope] в качестве Main dispatcher.
  */
